mod jsp
reglage de bug push

+++ SupPoke_Game/SupPoke_Game/SupPoke Game/include/dommage.h
@@ -2,7 +2,7 @@
 #define SUPPOKE_GAME_DOMMAGE_H
 
 int calculate_damage(struct move *move, struct supemon *launcher, struct supemon *target);
-float success_rate(struct move *move, struct supemon *launcher, struct supemon *target);
+int success_rate(struct move *move, struct supemon *launcher, struct supemon *target);
 int is_offensive(struct move *move);
 int is_passive(struct move *move);
 void execute_move(struct move *move, struct supemon *launcher, struct supemon *target);
